// packages/api-client/src/generated/client/types.gen.ts
export type ResponseStyle = 'data' | 'fields';

export interface Config<T extends ClientOptions = ClientOptions> extends Omit<RequestInit, 'body' | 'headers' | 'method'>, CoreConfig {
	/**
	 * Base URL for all requests made by this client.
	 */
	baseUrl?: T['baseUrl'];
	/**
	 * Fetch API implementation. You can use this option to provide a custom
	 * fetch instance.
	 *
	 * @default globalThis.fetch
	 */
	fetch?: typeof fetch;
	/**
	 * Please don't use the Fetch client for Next.js applications. The `next`
	 * options won't have any effect.
	 *
	 * Install {@link https://www.npmjs.com/package/@hey-api/client-next `@hey-api/client-next`} instead.
	 */
	next?: never;
	/**
	 * Return the response data parsed in a specified format. By default, `auto`
	 * will infer the appropriate method from the `Content-Type` response header.
	 * You can override this behavior with any of the {@link Body} methods.
	 * Select `stream` if you don't want to parse response data at all.
	 *
	 * @default 'auto'
	 */
	parseAs?: 'arrayBuffer' | 'auto' | 'blob' | 'formData' | 'json' | 'stream' | 'text';
	/**
	 * Should we return only data or multiple fields (data, error, response, etc.)?
	 *
	 * @default 'fields'
	 */
	responseStyle?: ResponseStyle;
	/**
	 * Throw an error instead of returning it in the response?
	 *
	 * @default false
	 */
	throwOnError?: T['throwOnError'];
}

export interface RequestOptions<
	TData = unknown,
	TResponseStyle extends ResponseStyle = 'fields',
	ThrowOnError extends boolean = boolean,
	Url extends string = string
>
	extends
// ...
export interface ResolvedRequestOptions<
	TResponseStyle extends ResponseStyle = 'fields',
	ThrowOnError extends boolean = boolean,
	Url extends string = string
> extends RequestOptions<unknown, TResponseStyle, ThrowOnError, Url> {
	headers: Headers;
	serializedBody?: string;
}

export type RequestResult<
	TData = unknown,
	TError = unknown,
	ThrowOnError extends boolean = boolean,
	TResponseStyle extends ResponseStyle = 'fields'
> = ThrowOnError extends true
	? Promise<
			TResponseStyle extends 'data'
				? TData extends Record<string, unknown>
					? TData[keyof TData]
					: TData
				: {
						data: TData extends Record<string, unknown> ? TData[keyof TData] : TData;
						request: Request;
						response: Response;
					}
		>
	: Promise<
			TResponseStyle extends 'data'
				? (TData extends Record<string, unknown> ? TData[keyof TData] : TData) | undefined
				: (
						| {
								data: TData extends Record<string, unknown> ? TData[keyof TData] : TData;
								error: undefined;
						  }
						| {
								data: undefined;
								error: TError extends Record<string, unknown> ? TError[keyof TError] : TError;
						  }
					) & {
						/** request may be undefined, because error may be from building the request object itself */
						request?: Request;
						/** response may be undefined, because error may be from building the request object itself or from a network error */
						response?: Response;
					}
		>;

export interface ClientOptions {
	baseUrl?: string;
	responseStyle?: ResponseStyle;
	throwOnError?: boolean;
}

type MethodFn = <TData = unknown, TError = unknown, ThrowOnError extends boolean = false, TResponseStyle extends ResponseStyle = 'fields'>(
	options: Omit<RequestOptions<TData, TResponseStyle, ThrowOnError>, 'method'>
) => RequestResult<TData, TError, ThrowOnError, TResponseStyle>;

type SseFn = <
	TData = unknown,
	// eslint-disable-next-line @typescript-eslint/no-unused-vars
	_TError = unknown,
	ThrowOnError extends boolean = false,
	TResponseStyle extends ResponseStyle = 'fields'
>(
	options: Omit<RequestOptions<never, TResponseStyle, ThrowOnError>, 'method'>
) => Promise<ServerSentEventsResult<TData>>;

type RequestFn = <TData = unknown, TError = unknown, ThrowOnError extends boolean = false, TResponseStyle extends ResponseStyle = 'fields'>(
	options: Omit<RequestOptions<TData, TResponseStyle, ThrowOnError>, 'method'> &
		Pick<Required<RequestOptions<TData, TResponseStyle, ThrowOnError>>, 'method'>
) => RequestResult<TData, TError, ThrowOnError, TResponseStyle>;

type BuildUrlFn = <
	TData extends {
		body?: unknown;
		path?: Record<string, unknown>;
		query?: Record<string, unknown>;
		url: string;
	}
>(
	options: TData & Options<TData>
) => string;

export type Client = CoreClient<RequestFn, Config, MethodFn, BuildUrlFn, SseFn> & {
	interceptors: Middleware<Request, Response, unknown, ResolvedRequestOptions>;
};
// ...
export type Options<
	TData extends TDataShape = TDataShape,
	ThrowOnError extends boolean = boolean,
	TResponse = unknown,
	TResponseStyle extends ResponseStyle = 'fields'
> = OmitKeys<RequestOptions<TResponse, TResponseStyle, ThrowOnError>, 'body' | 'path' | 'query' | 'url'> &
	([TData] extends [never] ? unknown : Omit<TData, 'url'>);